#86 Write a test for the definition of LiquibaseUtil
It is not possible to check the modifiers of a class using the PowerMockRunner, so I did a reqrite of the tests using JMockit. For JMockit working together with PowerMock, every test has to be run in its own JVM. That is slow but can be undone after PowerMock has been completely removed.